% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_species_data.R
\name{fetch_species_data}
\alias{fetch_species_data}
\title{fetch_species_data
The function checks scientific names status and returns several information around the names.}
\usage{
fetch_species_data(species)
}
\arguments{
\item{species}{a vector of species names}
}
\value{
a data.frame
}
\description{
fetch_species_data
The function checks scientific names status and returns several information around the names.
}
\examples{

library(httr)
library(jsonlite)
library(dplyr)
species_list <- c("Acetabularia parvula", "Codium decorticatum")
fetch_species_data(species_list)

}
