% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_valid_names.R
\name{get_valid_names}
\alias{get_valid_names}
\title{A wrapper of fetch species data
It is a convenient wrapper to map through a vector of names and give progress.}
\usage{
get_valid_names(species_list)
}
\arguments{
\item{species_list}{a vector of species names}
}
\value{
a data.frame
}
\description{
A wrapper of fetch species data
It is a convenient wrapper to map through a vector of names and give progress.
}
\examples{

species_list <- c("Acetabularia parvula", "Codium decorticatum")
get_valid_names(species_list)

}
