% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namechecker-package.R
\docType{package}
\name{namechecker-package}
\alias{namechecker}
\alias{namechecker-package}
\title{namechecker: Check Species Names Using Worms API}
\description{
The package is designed to seamlessly interface with the WoRMS (World Register of Marine Species) API, providing a straightforward method for verifying the acceptance of species names. While there are other packages available that offer functionalities centered around the WoRMS API, we found that many of them come with complex documentation and offer a multitude of tools that extend beyond our basic requirement of name validation. Our package aims to simplify this process by focusing solely on the essential task of checking the acceptance status of species names, making it more user-friendly and easier to navigate.
}
\author{
\strong{Maintainer}: Fabio Favoretto \email{ffavoretto@ucsd.edu} (\href{https://orcid.org/0000-0002-6493-4254}{ORCID})

}
\keyword{internal}
